+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BACKWARD INCOMPATIBILITY: DoDashApp.__init__() db_manager_kwargs default value changed from {} to {'enable_scenario_seq': True, 'future': True}
 ### Fixed
 - HomePageEdit.download_scenarios_callback: Fixed FutureWarning and close() warnings.
+- HomePageEdit.download_scenario_callback: Fixed FutureWarning and close() warnings.
 - DashApp and DoDashApp: mutable default arguments replaced
 ### Added
 - UserManual.md (in documentation folder)