+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]## [0.1.2.4b0]
+## [Unreleased]## [0.1.2.4b1]
+
+## [0.1.2.4b0]
 ### Changed
 - FoliumColumnVisualizationPage default map height increased from 800 to 1000
 - RunModelPage, do_model_class_drpdwn now automatically selects the first DoModelRunnerConfig in configs